+++ v6/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Version 6.0.0
 Major update (2026 release)!
+- New components: LS.Icons, LS.Multipane, LS.Layout
+- LS.Create now supports Emmet abbreviations! (eg. LS.Create("#myId>ul>li*3") instead of LS.Create("div", { id: "myId", ... }))
+- Added LS.Util.parseEmmet
+- [ ] Added UI effects
 - Added sanitize option to LS.Create
 - Optimized startup and reduced overhead
 - Added LS.Icons support
@@ -7,13 +11,14 @@ Major update (2026 release)!
 - LS now has it's own [icon library](https://github.com/the-lstv/ls-icons)!
 - Tabs: tab list can now be reordered by dragging with the reordableList option
 - Tabs: tabs.add now returns the tab ID
+- Enhancements to the default tab handle style
+- LS.Tabs change event is now change instead of changed
 - Deprecated the ambiguous LS.Color.autoScheme() in favor of LS.Color.autoSchemeEnabled
 - Renamed LS.LoadComponent -> LS.register, LS.GetComponent -> LS.getComponentByName, and LS.UnregisterComponent -> LS.unregisterComponent for consistency (methods starting with lowercase).
 - API changes: ESM export is now available (using the mjs extension), and browser compatibility level can be specified (?compat=legacy|modern|latest)
 - General improvements
-- LS.Tabs change event is now change instead of changed
 - EventEmitter.alias is now EventEmitter.aliasEvent for clarity
-- Multiple bugs fixed
+- Enhancements to LS.ImageCropper
 
 **Breaking changes:**
 - Removed legacy TinyFramework (LS.Tiny, Q, O, S, N, M). Now entirely under the LS namespace! (Temporarily, a "enableV5Compat" option exists that restores this.)